fio: use one file per job
It turns out that using multiple files per job causes issues with
distribution. There was no clear way to do zipf across the files, not
across the jobs.

We use a file per job make zipf across files possible.

Use 12 files and increase number of vCPUs to match that. 12 because each
NUMA node on seksu to which a VM is bound has 12 physical CPU cores.